+++ src/Paramecium/Paramecium/Forms/Renderer/SoupViewRenderer.cs
@@ -233,7 +233,7 @@ namespace Paramecium.Forms.Renderer
                                             DrawLine(
                                                 targetBitmap, targetGraphics, cameraPosition, cameraZoomFactor,
                                                 targetAnimal.Position + Double2d.FromAngle(targetAnimal.Angle + 0.5d) * 0.5d,
-                                                targetAnimal.Position + Double2d.FromAngle(targetAnimal.Angle + 0.5d) * 0.5d + Double2d.FromAngle(targetAnimal.Angle + 0.5d + (tailAngleRandom.NextDouble() * 2d - 1d) * 0.05d * targetAnimal.BrainOutput.Acceleration + double.Max(-1d, double.Min(1d, -targetAnimal.BrainOutput.Rotation)) * 0.15d) * (0.4d + tailLengthRandom.NextDouble() * 0.2d),
+                                                targetAnimal.Position + Double2d.FromAngle(targetAnimal.Angle + 0.5d) * 0.5d + Double2d.FromAngle(targetAnimal.Angle + 0.5d + (tailAngleRandom.NextDouble() * 2d - 1d) * 0.05d * double.Max(-1d, double.Min(1d, targetAnimal.BrainOutput.Acceleration)) + double.Max(-1d, double.Min(1d, -targetAnimal.BrainOutput.Rotation)) * 0.15d) * (0.4d + tailLengthRandom.NextDouble() * 0.2d),
                                                 Color.FromArgb(255, 255, 255)
                                             );
                                         }